refactor(Pagination):  add a basic IPageInfo type which represents the raw payload from API
