+++ packages/webcomponents/src/api/Pagination.ts
@@ -1,9 +1,12 @@
-export interface PagingInfo {
-  amount: number,
-  start_cursor: string,
-  end_cursor: string,
-  has_previous: boolean,
-  has_next: boolean,
+export interface IPageInfo {
+  start_cursor: string;
+  end_cursor: string;
+  has_previous: boolean;
+  has_next: boolean;
+}
+
+export interface PagingInfo extends IPageInfo {
+  amount: number;
 }
 
 export const pagingDefaults = {
@@ -12,7 +15,7 @@ export const pagingDefaults = {
   end_cursor: '',
   has_previous: false,
   has_next: false,
-}
+};
 
 export interface ExtendedPagingInfo extends PagingInfo {
   handleClickNext: (after_cursor: string) => void;
@@ -21,6 +24,6 @@ export interface ExtendedPagingInfo extends PagingInfo {
 
 export const ExtendedPagingDefaults: ExtendedPagingInfo = {
   ...pagingDefaults,
-  handleClickNext: () => { },
-  handleClickPrevious: () => { }
-}
+  handleClickNext: () => {},
+  handleClickPrevious: () => {},
+};
